Move NVI year filter to its own component
